requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
pandas
db-dtypes
sqlalchemy
pyodbc
python-dotenv